+++ motion_imitation/envs/sensors/space_utils.py
@@ -94,8 +94,8 @@ def convert_1d_box_sensors_to_gym_space(
 
   lower_bound = np.concatenate([s.get_lower_bound() for s in sensors])
   upper_bound = np.concatenate([s.get_upper_bound() for s in sensors])
-  observation_space = spaces.Box(np.array(lower_bound),
-                                 np.array(upper_bound),
+  observation_space = spaces.Box(np.array(lower_bound, dtype=np.float32),
+                                 np.array(upper_bound, dtype=np.float32),
                                  dtype=np.float32)
   return observation_space
 
@@ -117,8 +117,8 @@ def convert_sensors_to_gym_space_dictionary(
   gym_space_dict = {}
   for s in sensors:
     if isinstance(s, sensor.BoxSpaceSensor):
-      gym_space_dict[s.get_name()] = spaces.Box(np.array(s.get_lower_bound()),
-                                                np.array(s.get_upper_bound()),
+      gym_space_dict[s.get_name()] = spaces.Box(np.array(s.get_lower_bound(), dtype=np.float32),
+                                                np.array(s.get_upper_bound(), dtype=np.float32),
                                                 dtype=np.float32)
     else:
       raise UnsupportedConversionError('sensors = ' + str(sensors))
